+++ postpic/datahandling.py
@@ -349,7 +349,7 @@ class Axis(object):
 
     def __str__(self):
         s = '<Axis "{}" ({} grid points from {} to {})>'
-        return s.format(str(self.name), str(len(self)), *self.extent)
+        return s.format(self.name, len(self), *self.extent)
 
     __repr__ = __str__
 
@@ -1881,7 +1881,8 @@ class Field(NDArrayOperatorsMixin):
         self.export(filename)
 
     def __str__(self):
-        return '<Feld "' + self.name + '" ' + str(self.shape) + '>'
+        s = '<Field "{:}" {:}>'
+        return s.format(self.name, self.shape)
 
     def _extent_to_slices(self, extent):
         if not self.dimensions * 2 == len(extent):