+++ test/test_particles.py
@@ -45,5 +45,38 @@ class TestMultiSpecies(unittest.TestCase):
         lenc = len(p2)
         self.assertEqual(lenc, 3)
 
+    def test_repr(self):
+        print(self.p)
+        print(self.p._ssas[0])
+
+    def test_multifilter(self):
+        l0 = len(self.p)
+        s1 = self.p.filter('y>0')
+        l1 = len(s1)
+        s2 = s1.filter('x>0')
+        l2 = len(s2)
+        s3 = s2.filter('px>0')
+        l3 = len(s3)
+        self.assertEqual(len(self.p), l0)
+        self.assertEqual(len(s1), l1)
+        self.assertEqual(len(s2), l2)
+        self.assertEqual(len(s3), l3)
+
+    def test_invert(self):
+        l0 = len(self.p)
+        sgtr = self.p.filter('y>0')
+        lgtr = len(sgtr)
+        sless = self.p.filter('y<=0')
+        lless = len(sless)
+        self.assertEqual(lgtr + lless, l0)
+        self.assertEqual(lgtr, len(~sless))
+        self.assertEqual(lless, len(~sgtr))
+        self.assertEqual(lgtr, len(~~sgtr))
+
+    def test_invert2(self):
+        l0 = len(self.p)
+        self.assertEqual(len(~self.p), 0)
+        self.assertEqual(len(~~self.p), l0)
+
 if __name__ == '__main__':
     unittest.main()